Fix compiler errors in tests

+++ test/ReleasePackageTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Runtime.Versioning;
 using Squirrel.MarkdownSharp;
-using NuGet;
 using Squirrel;
 using Squirrel.Tests.TestHelpers;
 using System;
@@ -11,6 +10,7 @@ using System.Reflection;
 using System.Xml.Linq;
 using Squirrel.SimpleSplat;
 using Xunit;
+using Squirrel.NuGet;
 
 namespace Squirrel.Tests.Core
 {
@@ -30,13 +30,11 @@ namespace Squirrel.Tests.Core
         {
             var inputPackage = IntegrationTestHelper.GetPath("fixtures", "Squirrel.Tests.0.1.0-pre.nupkg");
             var outputPackage = Path.GetTempFileName() + ".nupkg";
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
 
             var fixture = new ReleasePackage(inputPackage);
-            (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
 
             try {
-                fixture.CreateReleasePackage(outputPackage, sourceDir);
+                fixture.CreateReleasePackage(outputPackage);
 
                 this.Log().Info("Resulting package is at {0}", outputPackage);
                 var pkg = new ZipPackage(outputPackage);
@@ -50,7 +48,7 @@ namespace Squirrel.Tests.Core
                 List<IPackageFile> files = pkg.GetFiles().ToList();
                 files.ForEach(x => this.Log().Info(x.Path));
 
-                List<string> nonDesktopPaths = new[] {"sl", "winrt", "netcore", "win8", "windows8", "MonoAndroid", "MonoTouch", "MonoMac", "wp", }
+                List<string> nonDesktopPaths = new[] { "sl", "winrt", "netcore", "win8", "windows8", "MonoAndroid", "MonoTouch", "MonoMac", "wp", }
                     .Select(x => @"lib\" + x)
                     .ToList();
 
@@ -62,82 +60,29 @@ namespace Squirrel.Tests.Core
         }
 
         [Fact]
-        public void FindPackageInOurLocalPackageList()
-        {
-            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "Squirrel.Core.1.0.0.0.nupkg");
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-            (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
-
-            var fixture = ExposedObject.From(new ReleasePackage(inputPackage));
-            IPackage result = fixture.matchPackage(new LocalPackageRepository(sourceDir), "xunit", VersionUtility.ParseVersionSpec("[1.0,2.0]"));
-
-            result.Id.ShouldEqual("xunit");
-            result.Version.Version.Major.ShouldEqual(2);
-            result.Version.Version.Minor.ShouldEqual(0);
-        }
-
-        [Fact]
-        public void FindDependentPackagesForDummyPackage()
-        {
-            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "Squirrel.Tests.0.1.0-pre.nupkg");
-            var fixture = new ReleasePackage(inputPackage);
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-            (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
-
-            IEnumerable<IPackage> results = fixture.findAllDependentPackages(default(IPackage), (IPackageRepository)new LocalPackageRepository(sourceDir), default(HashSet<string>), default(FrameworkName));
-            results.Count().ShouldBeGreaterThan(0);
-        }
-
-        [Fact]
         public void CanLoadPackageWhichHasNoDependencies()
         {
             var inputPackage = IntegrationTestHelper.GetPath("fixtures", "Squirrel.Core.NoDependencies.1.0.0.0.nupkg");
             var outputPackage = Path.GetTempFileName() + ".nupkg";
             var fixture = new ReleasePackage(inputPackage);
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
             try {
-                fixture.CreateReleasePackage(outputPackage, sourceDir);
-            }
-            finally {
+                fixture.CreateReleasePackage(outputPackage);
+            } finally {
                 File.Delete(outputPackage);
             }
         }
 
         [Fact]
-        public void CanResolveMultipleLevelsOfDependencies()
+        public void ThrowsIfLoadsPackageWithDependencies()
         {
-            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "Squirrel.Tests.0.1.0-pre.nupkg");
+            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "ProjectDependsOnJsonDotNet.1.0.nupkg");
             var outputPackage = Path.GetTempFileName() + ".nupkg";
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-
             var fixture = new ReleasePackage(inputPackage);
-            (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
-
             try {
-                fixture.CreateReleasePackage(outputPackage, sourceDir);
-
-                this.Log().Info("Resulting package is at {0}", outputPackage);
-                var pkg = new ZipPackage(outputPackage);
-
-                int refs = pkg.FrameworkAssemblies.Count();
-                this.Log().Info("Found {0} refs", refs);
-                refs.ShouldEqual(0);
-
-                this.Log().Info("Files in release package:");
-                pkg.GetFiles().ForEach(x => this.Log().Info(x.Path));
-
-                var filesToLookFor = new[] {
-                    "xunit.assert.dll",         // Tests => Xunit => Xunit.Assert
-                    "NuGet.Core.dll",           // Tests => NuGet
-                    "Squirrel.Tests.dll",
-                };
-
-                filesToLookFor.ForEach(name => {
-                    this.Log().Info("Looking for {0}", name);
-                    pkg.GetFiles().Any(y => y.Path.ToLowerInvariant().Contains(name.ToLowerInvariant())).ShouldBeTrue();
-                });
+                Assert.Throws<InvalidOperationException>(() => fixture.CreateReleasePackage(outputPackage));
             } finally {
-                File.Delete(outputPackage);
+                if (File.Exists(outputPackage))
+                    File.Delete(outputPackage);
             }
         }
 
@@ -159,7 +104,7 @@ namespace Squirrel.Tests.Core
                 // invulnerable to ExposedObject. Whyyyyyyyyy
                 var renderMinfo = fixture.GetType().GetMethod("renderReleaseNotesMarkdown",
                     BindingFlags.NonPublic | BindingFlags.Instance);
-                renderMinfo.Invoke(fixture, new object[] {targetFile, processor});
+                renderMinfo.Invoke(fixture, new object[] { targetFile, processor });
 
                 var doc = XDocument.Load(targetFile);
                 XNamespace ns = "http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd";
@@ -175,83 +120,14 @@ namespace Squirrel.Tests.Core
         }
 
         [Fact]
-        public void UsesTheRightVersionOfADependencyWhenMultipleAreInPackages()
-        {
-            var outputPackage = Path.GetTempFileName() + ".nupkg";
-            string outputFile = null;
-
-            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "CaliburnMicroDemo.1.0.0.nupkg");
-
-            var wrongPackage = "Caliburn.Micro.1.4.1.nupkg";
-            var wrongPackagePath = IntegrationTestHelper.GetPath("fixtures", wrongPackage);
-            var rightPackage = "Caliburn.Micro.1.5.2.nupkg";
-            var rightPackagePath = IntegrationTestHelper.GetPath("fixtures", rightPackage);
-
-            try {
-                var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-                (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
-
-                File.Copy(wrongPackagePath, Path.Combine(sourceDir, wrongPackage), true);
-                File.Copy(rightPackagePath, Path.Combine(sourceDir, rightPackage), true);
-
-                var package = new ReleasePackage(inputPackage);
-                var outputFileName = package.CreateReleasePackage(outputPackage, sourceDir);
-
-                var zipPackage = new ZipPackage(outputFileName);
-
-                var fileName = "Caliburn.Micro.dll";
-                var dependency = zipPackage.GetLibFiles()
-                    .Where(f => f.Path.EndsWith(fileName))
-                    .Single(f => f.TargetFramework == FrameworkTargetVersion.Net40);
-
-                outputFile = new FileInfo(Path.Combine(sourceDir, fileName)).FullName;
-
-                using (var of = File.Create(outputFile))
-                {
-                    dependency.GetStream().CopyTo(of);
-                }
-
-                var assemblyName = AssemblyName.GetAssemblyName(outputFile);
-                Assert.Equal(1, assemblyName.Version.Major);
-                Assert.Equal(5, assemblyName.Version.Minor);
-            } finally {
-                File.Delete(outputPackage);
-                File.Delete(outputFile);
-            }
-        }
-
-        [Fact]
-        public void DependentPackageNotFoundAndThrowsError()
-        {
-            string packagesDir;
-            // use empty packages folder
-            using (Utility.WithTempDirectory(out packagesDir)) {
-                var inputPackage = IntegrationTestHelper.GetPath("fixtures", "ProjectDependsOnJsonDotNet.1.0.nupkg");
-
-                var outputPackage = Path.GetTempFileName() + ".nupkg";
-
-                try {
-                    var package = new ReleasePackage(inputPackage);
-                    Assert.Throws<Exception>(() =>
-                        package.CreateReleasePackage(outputPackage, packagesDir));
-                } finally {
-                    File.Delete(outputPackage);
-                }
-            }
-        }
-
-        [Fact]
         public void ContentFilesAreIncludedInCreatedPackage()
         {
             var inputPackage = IntegrationTestHelper.GetPath("fixtures", "ProjectWithContent.1.0.0.0-beta.nupkg");
             var outputPackage = Path.GetTempFileName() + ".zip";
-            var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-
             var fixture = new ReleasePackage(inputPackage);
-            (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
 
             try {
-                fixture.CreateReleasePackage(outputPackage, sourceDir);
+                fixture.CreateReleasePackage(outputPackage);
 
                 this.Log().Info("Resulting package is at {0}", outputPackage);
                 var pkg = new ZipPackage(outputPackage);
@@ -271,37 +147,6 @@ namespace Squirrel.Tests.Core
                 Assert.Contains("dir\\item-in-subdirectory.txt", contentFilePaths);
 
                 Assert.Equal(1, pkg.GetLibFiles().Count());
-            } finally {
-                File.Delete(outputPackage);
-            }
-        }
-
-        [Fact]
-        public void WhenAProjectContainsNet45BinariesItContainsTheNecessaryDependency()
-        {
-            var outputPackage = Path.GetTempFileName() + ".nupkg";
-
-            var inputPackage = IntegrationTestHelper.GetPath("fixtures", "ThisShouldBeANet45Project.1.0.nupkg");
-
-            var rightPackage = "Caliburn.Micro.1.5.2.nupkg";
-            var rightPackagePath = IntegrationTestHelper.GetPath("fixtures", rightPackage);
-
-            try {
-                var sourceDir = IntegrationTestHelper.GetPath("fixtures", "packages");
-                (new DirectoryInfo(sourceDir)).Exists.ShouldBeTrue();
-
-                File.Copy(rightPackagePath, Path.Combine(sourceDir, rightPackage), true);
-
-                var package = new ReleasePackage(inputPackage);
-                var outputFileName = package.CreateReleasePackage(outputPackage, sourceDir);
-
-                var zipPackage = new ZipPackage(outputFileName);
-
-                var dependency = zipPackage.GetLibFiles()
-                    .Where(f => f.Path.EndsWith("Caliburn.Micro.dll"))
-                    .FirstOrDefault(f => f.TargetFramework == FrameworkTargetVersion.Net45);
-
-                Assert.NotNull(dependency);
             } finally {
                 File.Delete(outputPackage);
             }
